@@ -6,10 +6,10 @@
     "dev": "vite"
   },
   "dependencies": {
-    "vue": "2.7.0-beta.7"
+    "vue": "3.2.37"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue2": "^1.1.0",
+    "@vitejs/plugin-vue": "^2.3.3",
     "vite": "^2.9.13"
   }
 }
