@@ -1,4 +1,4 @@
-import Vue from 'vue'
+import {createApp} from 'vue'
 import App from './App.vue'
 
-new Vue({...App, el: '#app'})
+createApp(App).mount('#app')
